FAST_API/requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
groq
colorama
python-dotenv
typing-extensions
starlette